:recycle: Novel 평점 실수형으로 변경 - 평점을 소수점 첫째자리 까지 반환하도록 변경 - review가 달리지않은 작품의 경우 null을 반환합니다.

+++ src/main/java/com/yju/toonovel/domain/novel/dto/LikeNovelPaginationResponseDto.java
@@ -14,12 +14,12 @@ public class LikeNovelPaginationResponseDto {
 	private String title;
 	private String author;
 	private Genre genre;
-	private Integer grade;
+	private Double grade;
 	private String image;
 
 	@Builder
 	public LikeNovelPaginationResponseDto(Long likeNovelId, Long novelId, String title, String author, Genre genre,
-		Integer grade, String image) {
+		Double grade, String image) {
 		this.likeNovelId = likeNovelId;
 		this.novelId = novelId;
 		this.title = title;
